+++ src/canister/src/lib.rs
@@ -19,7 +19,6 @@ pub fn __export_did_tmp_() -> String {
     use candid::export_service;
     use canister_types::models::api_error::*;
     use canister_types::models::attendee::*;
-    use canister_types::models::boosted::*;
     use canister_types::models::event::*;
     use canister_types::models::filter_type::*;
     use canister_types::models::friend_request::*;